requirements.txt
Django==1.4
distribute==0.6.19
django-registration==0.8
docutils==0.9
wsgiref==0.1.2
